Added support for verbose alternative
