+++ README.md
@@ -1,17 +1,13 @@
 # assert_tokens_eq
 
-Like Rust's built-in `assert_eq` macro, but for token streams.
-Passes them through rustfmt,
-and shows a pretty diff
-(powered by [pretty_assertions]).
-
-[pretty_assertions]: https://crates.io/crates/pretty_assertions
+Like Rust's built-in [`assert_eq`] macro, but for token streams.
+Passes them through rustfmt, and shows a pretty diff (powered by [`pretty_assertions`]).
 
 ## How it works
 
 You write:
 
-```rust
+```rust, ignore
 let got = something_that_generates_rust_code();
 let expected = quote! {
     #[no_mangle]
@@ -30,14 +26,14 @@ assert_tokens_eq!(got, expected);
 
 and you get:
 
-![](./screenshot.png)
+![Screenshot](./screenshot.png)
 
 ## License
 
 Licensed under either of
 
- * Apache License, Version 2.0, ([LICENSE-APACHE](LICENSE-APACHE) or http://www.apache.org/licenses/LICENSE-2.0)
- * MIT license ([LICENSE-MIT](LICENSE-MIT) or http://opensource.org/licenses/MIT)
+* Apache License, Version 2.0, ([LICENSE-APACHE](LICENSE-APACHE) or <http://www.apache.org/licenses/LICENSE-2.0>)
+* MIT license ([LICENSE-MIT](LICENSE-MIT) or <http://opensource.org/licenses/MIT>)
 
 at your option.
 
@@ -47,3 +43,6 @@ Unless you explicitly state otherwise, any contribution intentionally
 submitted for inclusion in the work by you, as defined in the Apache-2.0
 license, shall be dual licensed as above, without any additional terms or
 conditions.
+
+[`assert_eq`]: https://doc.rust-lang.org/std/macro.assert_eq.html
+[`pretty_assertions`]: https://crates.io/crates/pretty_assertions
